+++ src/components/MessageBox.vue
@@ -67,8 +67,8 @@
     import { Chat, Message, MessageBodyType, MessageTypes } from '@/types';
     import SelectedOptions from '@/components/fileBrowser/OptionsSelectedItems.vue';
     import Dialog from '@/components/Dialog.vue';
-    import EditShare from '@/components/fileBrowser/EditShare.vue';
-    import ShareChatTable from '@/components/fileBrowser/ShareChatTable.vue';
+    import EditShare from '@/components/fileBrowser/OptionsSelectedItem/EditShare.vue';
+    import ShareChatTable from '@/components/fileBrowser/OptionsSelectedItem/ShareChatTable.vue';
 
     interface IProps {
         chat: Chat;
