+++ src/components/fileBrowser/OptionsSelectedItem/EditShare.vue
