+++ src/components/fileBrowser/OptionsSelectedItem/ShareChatTable.vue
